Fix column calculation and begin test level 3: table with different row columns.

+++ HTMLParser/HTMLFileParser.cs
@@ -2,6 +2,7 @@ using System;
 using System.Xml.XPath;
 using System.Linq;
 using System.IO;
+using System.Text;
 using HtmlAgilityPack;
 
 namespace HTMLParser
@@ -23,33 +24,39 @@ namespace HTMLParser
 
 		public bool Open(string filePath) {
 			try {
-				_htmlDoc = new HtmlDocument();
+                if (File.Exists(filePath))
+                {
+                    _htmlDoc = new HtmlDocument();
 
-				// There are various options, set as needed
-				_htmlDoc.OptionFixNestedTags = true;
-				_htmlDoc.OptionOutputOriginalCase = true;
-				_htmlDoc.OptionWriteEmptyNodes = true;
+                    // There are various options, set as needed
+                    _htmlDoc.OptionFixNestedTags = true;
+                    _htmlDoc.OptionOutputOriginalCase = true;
+                    _htmlDoc.OptionWriteEmptyNodes = true;
 
-				// filePath is a path to a file containing the html
-				_htmlDoc.Load(filePath);
-				// Use:  htmlDoc.LoadHtml(xmlString);  to load from a string (was htmlDoc.LoadXML(xmlString)
+                    // filePath is a path to a file containing the html
+                    _htmlDoc.Load(filePath);
+                    // Use:  htmlDoc.LoadHtml(xmlString);  to load from a string (was htmlDoc.LoadXML(xmlString)
 
-				// ParseErrors is an ArrayList containing any errors from the Load statement
-				return !(_htmlDoc.ParseErrors != null && _htmlDoc.ParseErrors.Count() > 0);
-
+                    // ParseErrors is an ArrayList containing any errors from the Load statement
+                    return !(_htmlDoc.ParseErrors != null && _htmlDoc.ParseErrors.Count() > 0);
+                } else {
+                    return false;
+                }
 			} catch {
 				return false;
 			}
 
 		}
 
-		public void Describe() {
+		public void Describe(string fileName) {
+            StringBuilder sb = new StringBuilder();
 			if (_htmlDoc != null) {
 				using (HTMLElement rootElement = new HTMLElement()) {
 					rootElement.AppendNode(_htmlDoc.DocumentNode.ChildNodes.FindFirst(HTMLElement.HtmlRootTag));
-					rootElement.Describe();
+					rootElement.Describe(sb);
 				}
 			}
+            File.WriteAllText(fileName, sb.ToString());
 		}
 
 		public bool Fix() {
@@ -75,7 +82,9 @@ namespace HTMLParser
 					if (File.Exists(backupFilePath)) {
 						File.Delete(backupFilePath);
 					}
-					File.Copy(FilePath, backupFilePath);
+                    if (File.Exists(FilePath)) {
+                        File.Copy(FilePath, backupFilePath);
+                    }
 				}
 				_htmlDoc.Save(FilePath);
 				return true;
